Put components in components folder
