+++ consistent-evaluation.js
@@ -1,56 +1,1 @@
-import { css, html, LitElement } from 'lit-element/lit-element.js';
-import { LocalizeMixin } from '@brightspace-ui/core/mixins/localize-mixin.js';
-
-class ConsistentEvaluation extends LocalizeMixin(LitElement) {
-
-	static get properties() {
-		return {
-			prop1: { type: String },
-		};
-	}
-
-	static get styles() {
-		return css`
-			:host {
-				display: inline-block;
-			}
-			:host([hidden]) {
-				display: none;
-			}
-		`;
-	}
-
-	static async getLocalizeResources(langs) {
-		for await (const lang of langs) {
-			let translations;
-			switch (lang) {
-				case 'en':
-					translations = await import('./locales/en.js');
-					break;
-			}
-
-			if (translations && translations.val) {
-				return {
-					language: lang,
-					resources: translations.val
-				};
-			}
-		}
-
-		return null;
-	}
-
-	constructor() {
-		super();
-
-		this.prop1 = 'consistent-evaluation';
-	}
-
-	render() {
-		return html`
-			<h2>Hello ${this.prop1}!</h2>
-			<div>Localization Example: ${this.localize('myLangTerm')}</div>
-		`;
-	}
-}
-customElements.define('d2l-consistent-evaluation', ConsistentEvaluation);
+import './components/consistent-evaluation.js';
